requirements.txt
# Face Recognition Attendance System - Requirements

# Core dependencies
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# GUI framework
tk>=0.1.0

# Additional utilities
python-dateutil>=2.8.2
